move `ErrorController` into the `Errors` namespace
