+++ routes/cp.php
@@ -23,7 +23,7 @@ Route::prefix('seo-pro')->name('seo-pro.')->group(function () {
     Route::get('section-defaults/taxonomies/{seo_pro_taxonomy}/edit', [Controllers\CP\TaxonomyDefaultsController::class, 'edit'])->name('section-defaults.taxonomies.edit');
     Route::patch('section-defaults/taxonomies/{seo_pro_taxonomy}', [Controllers\CP\TaxonomyDefaultsController::class, 'update'])->name('section-defaults.taxonomies.update');
 
-    Route::resource('errors', Controllers\CP\ErrorController::class)->only('index');
+    Route::resource('errors', Controllers\CP\Errors\ErrorController::class)->only('index');
 
     Route::get('redirects/export', Controllers\CP\Redirects\ExportRedirectsController::class)->name('redirects.export');
     Route::post('redirects/import', Controllers\CP\Redirects\ImportRedirectsController::class)->name('redirects.import');
